Implement command handling using dynamic imports
The main file `sfit.py` imports subcommand modules based on a list
of subcommands in the main file. Each subcommand module initializes
its own parsers and attaches the handler function to the parsed
arguments.

This will allow for isolation of the declaration as well as
implementation of each subcommand. This makes it easier to add
new commands and keeps the main file clutter-free.