Add HEAD object file and Command line Interface file
I have added the new HEAD object file and the implementations of the commands related to the HEAD object. I have currently left the `command_handler` blank and will implement it so that it's consistent with the others.
Extra commands also can be added, I'm just waiting for now

Note: We also need to intialize the HEAD object (file) during the repo creation (init file).
That will have a path of refs/heads/main by default.